#376 Move Thought Archive components into new directories
