+++ ui/src/react/pages/archives/ArchivesPageWrapper.tsx
@@ -39,10 +39,10 @@ const containerElementName = 'archivesPageWrapper';
   template: `<span #${containerElementName}></span>`,
   styleUrls: [
     './ArchivesPage.scss',
-    './thought-archives/ArchivedBoardsList.scss',
-    './thought-archives/ArchivedBoardTile.scss',
-    './thought-archives/ArchivedBoardColumn.scss',
-    './thought-archives/ArchivedBoard.scss',
+    './thought-archives/archived-boards-list/ArchivedBoardsList.scss',
+    './thought-archives/archived-boards-list/archived-board-tile/ArchivedBoardTile.scss',
+    './thought-archives/archived-board/ArchivedBoard.scss',
+    './thought-archives/archived-board/archived-board-column/ArchivedBoardColumn.scss',
     './archives-subheader/ArchivesSubheader.scss',
     './action-item-archives/ActionItemArchives.scss',
     '../../components/column-header/ColumnHeader.scss',
