+++ ui/src/react/pages/archives/thought-archives/archived-board/ArchivedBoard.scss
@@ -1,4 +1,4 @@
-@import '../styles/mixins';
+@import '../../../../../styles/mixins';
 
 .archived-board {
   display: flex;
