+++ ui/src/react/pages/archives/thought-archives/archived-board/archived-board-column/ArchivedBoardColumn.scss
@@ -1,5 +1,5 @@
-@import '../styles/colors';
-@import '../styles/mixins';
+@import '../../../../../../styles/colors';
+@import '../../../../../../styles/mixins';
 
 .archived-column {
   width: 100%;
