+++ ui/src/react/pages/archives/thought-archives/archived-boards-list/ArchivedBoardsList.scss
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-@import '../styles/mixins';
-@import '../styles/colors';
+@import '../../../../../styles/mixins';
+@import '../../../../../styles/colors';
 
 .archived-boards-list {
   text-align: center;
