+++ ui/src/react/pages/archives/thought-archives/archived-boards-list/archived-board-tile/ArchivedBoardTile.scss
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-@import '../styles/mixins';
-@import '../styles/colors';
+@import '../../../../../../styles/mixins';
+@import '../../../../../../styles/colors';
 
 .archived-board-tile {
   display: flex;
