+++ ui/src/react/pages/archives/thought-archives/archived-boards-list/archived-board-tile/ArchivedBoardTile.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 
-import { mockBoards } from '../../../services/api/__mocks__/BoardService';
+import { mockBoards } from '../../../../../services/api/__mocks__/BoardService';
 
 import ArchivedBoardTile from './ArchivedBoardTile';
